requirements.txt
Pillow==8.4.0
diffusers==0.3.0
